# Conflicts:
#   languages/ja.json

+++ libs/uploaders/amazonS3.js
@@ -1,12 +1,12 @@
 var fs = require('fs');
 module.exports = function(s,config,lang){
     //Amazon S3
-    var beforeAccountSaveForAmazonS3 = function(d){
+    function beforeAccountSave(d){
         //d = save event
         d.formDetails.aws_use_global=d.d.aws_use_global
         d.formDetails.use_aws_s3=d.d.use_aws_s3
     }
-    var cloudDiskUseStartupForAmazonS3 = function(group,userDetails){
+    function cloudDiskUseStartup(group,userDetails){
         group.cloudDiskUse['s3'].name = 'Amazon S3'
         group.cloudDiskUse['s3'].sizeLimitCheck = (userDetails.use_aws_s3_size_limit === '1')
         if(!userDetails.aws_s3_size_limit || userDetails.aws_s3_size_limit === ''){
@@ -15,7 +15,7 @@ module.exports = function(s,config,lang){
             group.cloudDiskUse['s3'].sizeLimit = parseFloat(userDetails.aws_s3_size_limit)
         }
     }
-    var loadAmazonS3ForUser = function(e){
+    function loadGroupApp(e){
         // e = user
         var userDetails = JSON.parse(e.details)
         if(userDetails.aws_use_global === '1' && config.cloudUploaders && config.cloudUploaders.AmazonS3){
@@ -54,11 +54,11 @@ module.exports = function(s,config,lang){
             s.group[e.ke].aws_s3 = new s.group[e.ke].aws.S3();
         }
     }
-    var unloadAmazonS3ForUser = function(user){
+    function unloadGroupApp(user){
         s.group[user.ke].aws = null
         s.group[user.ke].aws_s3 = null
     }
-    var deleteVideoFromAmazonS3 = function(e,video,callback){
+    function deleteVideo(e,video,callback){
         // e = user
         try{
             var videoDetails = JSON.parse(video.details)
@@ -68,6 +68,9 @@ module.exports = function(s,config,lang){
         if(!videoDetails.location){
             videoDetails.location = video.href.split('.amazonaws.com')[1]
         }
+        if(videoDetails.type !== 's3'){
+            return
+        }
         s.group[e.ke].aws_s3.deleteObject({
             Bucket: s.group[e.ke].init.aws_s3_bucket,
             Key: videoDetails.location,
@@ -76,7 +79,7 @@ module.exports = function(s,config,lang){
             callback()
         });
     }
-    var uploadVideoToAmazonS3 = function(e,k){
+    function uploadVideo(e,k){
         //e = video object
         //k = temporary values
         if(!k)k={};
@@ -92,9 +95,8 @@ module.exports = function(s,config,lang){
             s.group[e.ke].aws_s3.upload({
                 Bucket: s.group[e.ke].init.aws_s3_bucket,
                 Key: saveLocation,
-                Body:fileStream,
-                ACL:'public-read',
-                ContentType:'video/'+ext
+                Body: fileStream,
+                ContentType: 'video/'+ext
             },function(err,data){
                 if(err){
                     s.userLog(e,{type:lang['Amazon S3 Upload Error'],msg:err})
@@ -114,7 +116,7 @@ module.exports = function(s,config,lang){
                             }),
                             size: k.filesize,
                             end: k.endTime,
-                            href: data.Location
+                            href: ''
                         }
                     })
                     s.setCloudDiskUsedForGroup(e.ke,{
@@ -126,7 +128,7 @@ module.exports = function(s,config,lang){
             })
         }
     }
-    var onInsertTimelapseFrame = function(monitorObject,queryInfo,filePath){
+    function onInsertTimelapseFrame(monitorObject,queryInfo,filePath){
         var e = monitorObject
         if(s.group[e.ke].aws_s3 && s.group[e.ke].init.use_aws_s3 !== '0' && s.group[e.ke].init.aws_s3_save === '1'){
             var fileStream = fs.createReadStream(filePath)
@@ -152,6 +154,7 @@ module.exports = function(s,config,lang){
                             mid: queryInfo.mid,
                             ke: queryInfo.ke,
                             time: queryInfo.time,
+                            filename: queryInfo.filename,
                             details: s.s({
                                 type : 's3',
                                 location : saveLocation
@@ -169,7 +172,7 @@ module.exports = function(s,config,lang){
             })
         }
     }
-    var onDeleteTimelapseFrameFromCloud = function(e,frame,callback){
+    function onDeleteTimelapseFrameFromCloud(e,frame,callback){
         // e = user
         try{
             var frameDetails = JSON.parse(frame.details)
@@ -190,18 +193,30 @@ module.exports = function(s,config,lang){
             callback()
         });
     }
+    function onGetVideoData(video){
+        const videoDetails = s.parseJSON(video.details)
+        return new Promise((resolve, reject) => {
+            const saveLocation = videoDetails.location
+            var fileStream = s.group[video.ke].aws_s3.getObject({
+                Bucket: s.group[video.ke].init.aws_s3_bucket,
+                Key: saveLocation,
+            }).createReadStream();
+            resolve(fileStream)
+        })
+    }
     //amazon s3
     s.addCloudUploader({
         name: 's3',
-        loadGroupAppExtender: loadAmazonS3ForUser,
-        unloadGroupAppExtender: unloadAmazonS3ForUser,
-        insertCompletedVideoExtender: uploadVideoToAmazonS3,
-        deleteVideoFromCloudExtensions: deleteVideoFromAmazonS3,
-        cloudDiskUseStartupExtensions: cloudDiskUseStartupForAmazonS3,
-        beforeAccountSave: beforeAccountSaveForAmazonS3,
-        onAccountSave: cloudDiskUseStartupForAmazonS3,
+        loadGroupAppExtender: loadGroupApp,
+        unloadGroupAppExtender: unloadGroupApp,
+        insertCompletedVideoExtender: uploadVideo,
+        deleteVideoFromCloudExtensions: deleteVideo,
+        cloudDiskUseStartupExtensions: cloudDiskUseStartup,
+        beforeAccountSave: beforeAccountSave,
+        onAccountSave: cloudDiskUseStartup,
         onInsertTimelapseFrame: onInsertTimelapseFrame,
-        onDeleteTimelapseFrameFromCloud: onDeleteTimelapseFrameFromCloud
+        onDeleteTimelapseFrameFromCloud: onDeleteTimelapseFrameFromCloud,
+        onGetVideoData
     })
     //return fields that will appear in settings
     return {
